+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-### Feature 
+### Feature
 * Eth2 Staking Feature [#2662](https://github.com/MyEtherWallet/MyEtherWallet/pull/2662)
+
+### Bug
+
+* Fix error on print [#2676](https://github.com/MyEtherWallet/MyEtherWallet/pull/2676)
 
 ### Release v5.7.14